Changed type of get phone users endpoint response
