+++ Source/ZoomNet/Json/ZoomNetJsonSerializerContext.cs
@@ -149,7 +149,6 @@ namespace ZoomNet.Json
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallRecordingType))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUser))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUserProfile))]
-	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUserProfilesPaginationObject))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUserStatus))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneNumber))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneType))]
@@ -446,7 +445,6 @@ namespace ZoomNet.Json
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallRecordingType[]))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUser[]))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUserProfile[]))]
-	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUserProfilesPaginationObject[]))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneCallUserStatus[]))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneNumber[]))]
 	[JsonSerializable(typeof(ZoomNet.Models.PhoneType[]))]
